Added: OCRRecognizer class and baseline specification
OCRRecognizer is to be the basis of the BankOCR system.  There is a
higher level class which is going to represent potential bank account
numbers, but OCRRecognizer is going to have the responsibility of
turning input strings into recognised characters where possible, and
glyphs where recognition does not provide an exact match.

OCRRecognizer is to be an immutable object - you cannot change its
state after creation.  This therefore represents a single attempt at
recognition.