Added: Expectation for convenience constructor and initial input
OCRRecognizer is going to have a convenience constructor that takes a
character set to match against, and the input strings to match.
OCRCharacterSet is going to be a collection of recognisable characters
and the glyphs that they match.

This commit is enough to get the code to compile and the existence test
to pass.