Create a failing test for recognizer length
We should expect the recognizer to find a certain number of characters
from the input - they don't need to be valid characters, but characters
none the less.  We set up the expectation for this and stub to satisfy
the compiler.  The testing is failing, because without parsing some
input we can't meet the expectation.